requirements.txt
# YouTube Shorts Automation - Dependências
# Versões compatíveis com Python 3.12+

# Processamento de vídeo
moviepy>=2.0.0
opencv-python>=4.8.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8

# Download de vídeos
yt-dlp>=2023.12.30

# YouTube API
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Análise de áudio e fala
SpeechRecognition>=3.10.0
pydub>=0.25.1

# Processamento numérico
numpy>=1.24.0

# Agendamento e sistema
schedule>=1.2.0
psutil>=5.9.0

# Utilitários
requests>=2.31.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
Pillow>=10.0.0

# Ferramentas de build
setuptools>=65.0.0
wheel>=0.38.0